requirements.txt
Django>=5.0.2,<5.1
psycopg2>=2.9.9,<3.0
import-export==0.2.67.dev6
django-import-export #==2.7.1
pillow>= 10.2.0,<11.0
djangorestframework>=3.15.2,<4.0
django-cors-headers>=4.0.0,<5.0
python-dateutil==2.9.0.post0
django-filter==25.1
drf-yasg==1.21.10